requirements.txt
pyautogen==0.2.32
openai==1.35.0
pytest>=7.0.0
